Vertex/EdgeSeq, __plot__, operators
