+++ src/igraph/__init__.py
@@ -132,6 +132,8 @@ from igraph.drawing.colors import (
     palettes,
     known_colors,
 )
+from igraph.drawing.graph import __plot__ as _graph_plot
+from igraph.drawing.utils import autocurve
 from igraph.datatypes import Matrix, DyadCensus, TriadCensus, UniqueIdGenerator
 from igraph.formula import construct_graph_from_formula
 from igraph.io.files import (
@@ -183,7 +185,13 @@ from igraph.io.bipartite import (
 from igraph.io.images import _write_graph_to_svg
 from igraph.layout import Layout
 from igraph.matching import Matching
-from igraph.operators import disjoint_union, union, intersection
+from igraph.operators import (
+    disjoint_union,
+    union,
+    intersection,
+    operator_method_registry as _operator_method_registry,
+)
+from igraph.seq import EdgeSeq, VertexSeq
 from igraph.statistics import (
     FittedPowerLaw,
     Histogram,
@@ -1736,183 +1744,23 @@ class Graph(GraphBase):
 
     ###################
     # Custom operators
+    __iadd__ = _operator_method_registry['__iadd__']
 
-    def __iadd__(self, other):
-        """In-place addition (disjoint union).
+    __add__ = _operator_method_registry['__add__']
 
-        @see: L{__add__}
-        """
-        if isinstance(other, (int, str)):
-            self.add_vertices(other)
-            return self
-        elif isinstance(other, tuple) and len(other) == 2:
-            self.add_edges([other])
-            return self
-        elif isinstance(other, list):
-            if not other:
-                return self
-            if isinstance(other[0], tuple):
-                self.add_edges(other)
-                return self
-            if isinstance(other[0], str):
-                self.add_vertices(other)
-                return self
-        return NotImplemented
+    __and__ = _operator_method_registry['__and__']
 
-    def __add__(self, other):
-        """Copies the graph and extends the copy depending on the type of
-        the other object given.
+    __isub__ = _operator_method_registry['__isub__']
 
-        @param other: if it is an integer, the copy is extended by the given
-          number of vertices. If it is a string, the copy is extended by a
-          single vertex whose C{name} attribute will be equal to the given
-          string. If it is a tuple with two elements, the copy
-          is extended by a single edge. If it is a list of tuples, the copy
-          is extended by multiple edges. If it is a L{Graph}, a disjoint
-          union is performed.
-        """
-        if isinstance(other, (int, str)):
-            g = self.copy()
-            g.add_vertices(other)
-        elif isinstance(other, tuple) and len(other) == 2:
-            g = self.copy()
-            g.add_edges([other])
-        elif isinstance(other, list):
-            if len(other) > 0:
-                if isinstance(other[0], tuple):
-                    g = self.copy()
-                    g.add_edges(other)
-                elif isinstance(other[0], str):
-                    g = self.copy()
-                    g.add_vertices(other)
-                elif isinstance(other[0], Graph):
-                    return self.disjoint_union(other)
-                else:
-                    return NotImplemented
-            else:
-                return self.copy()
+    __sub__ = _operator_method_registry['__sub__']
 
-        elif isinstance(other, Graph):
-            return self.disjoint_union(other)
-        else:
-            return NotImplemented
+    __mul__ = _operator_method_registry['__mul__']
 
-        return g
-
-    def __and__(self, other):
-        """Graph intersection operator.
-
-        @param other: the other graph to take the intersection with.
-        @return: the intersected graph.
-        """
-        if isinstance(other, Graph):
-            return self.intersection(other)
-        else:
-            return NotImplemented
-
-    def __isub__(self, other):
-        """In-place subtraction (difference).
-
-        @see: L{__sub__}"""
-        if isinstance(other, int):
-            self.delete_vertices([other])
-        elif isinstance(other, tuple) and len(other) == 2:
-            self.delete_edges([other])
-        elif isinstance(other, list):
-            if len(other) > 0:
-                if isinstance(other[0], tuple):
-                    self.delete_edges(other)
-                elif isinstance(other[0], (int, str)):
-                    self.delete_vertices(other)
-                else:
-                    return NotImplemented
-        elif isinstance(other, Vertex):
-            self.delete_vertices(other)
-        elif isinstance(other, VertexSeq):
-            self.delete_vertices(other)
-        elif isinstance(other, Edge):
-            self.delete_edges(other)
-        elif isinstance(other, EdgeSeq):
-            self.delete_edges(other)
-        else:
-            return NotImplemented
-        return self
-
-    def __sub__(self, other):
-        """Removes the given object(s) from the graph
-
-        @param other: if it is an integer, removes the vertex with the given
-          ID from the graph (note that the remaining vertices will get
-          re-indexed!). If it is a tuple, removes the given edge. If it is
-          a graph, takes the difference of the two graphs. Accepts
-          lists of integers or lists of tuples as well, but they can't be
-          mixed! Also accepts L{Edge} and L{EdgeSeq} objects.
-        """
-        if isinstance(other, Graph):
-            return self.difference(other)
-
-        result = self.copy()
-        if isinstance(other, (int, str)):
-            result.delete_vertices([other])
-        elif isinstance(other, tuple) and len(other) == 2:
-            result.delete_edges([other])
-        elif isinstance(other, list):
-            if len(other) > 0:
-                if isinstance(other[0], tuple):
-                    result.delete_edges(other)
-                elif isinstance(other[0], (int, str)):
-                    result.delete_vertices(other)
-                else:
-                    return NotImplemented
-            else:
-                return result
-        elif isinstance(other, Vertex):
-            result.delete_vertices(other)
-        elif isinstance(other, VertexSeq):
-            result.delete_vertices(other)
-        elif isinstance(other, Edge):
-            result.delete_edges(other)
-        elif isinstance(other, EdgeSeq):
-            result.delete_edges(other)
-        else:
-            return NotImplemented
-
-        return result
-
-    def __mul__(self, other):
-        """Copies exact replicas of the original graph an arbitrary number of
-        times.
-
-        @param other: if it is an integer, multiplies the graph by creating the
-          given number of identical copies and taking the disjoint union of
-          them.
-        """
-        if isinstance(other, int):
-            if other == 0:
-                return Graph()
-            elif other == 1:
-                return self
-            elif other > 1:
-                return self.disjoint_union([self] * (other - 1))
-            else:
-                return NotImplemented
-
-        return NotImplemented
+    __or__ = _operator_method_registry['__or__']
 
     def __bool__(self):
         """Returns True if the graph has at least one vertex, False otherwise."""
         return self.vcount() > 0
-
-    def __or__(self, other):
-        """Graph union operator.
-
-        @param other: the other graph to take the union with.
-        @return: the union graph.
-        """
-        if isinstance(other, Graph):
-            return self.union(other)
-        else:
-            return NotImplemented
 
     def __coerce__(self, other):
         """Coercion rules.
@@ -1965,204 +1813,7 @@ class Graph(GraphBase):
     __iter__ = None  # needed for PyPy
     __hash__ = None  # needed for PyPy
 
-    def __plot__(self, backend, context, *args, **kwds):
-        """Plots the graph to the given Cairo context or matplotlib Axes.
-
-        The visual style of vertices and edges can be modified at three
-        places in the following order of precedence (lower indices override
-        higher indices):
-
-          1. Keyword arguments of this function (or of L{plot()} which is
-             passed intact to C{Graph.__plot__()}.
-
-          2. Vertex or edge attributes, specified later in the list of
-             keyword arguments.
-
-          3. igraph-wide plotting defaults (see
-             L{igraph.config.Configuration})
-
-          4. Built-in defaults.
-
-        E.g., if the C{vertex_size} keyword attribute is not present,
-        but there exists a vertex attribute named C{size}, the sizes of
-        the vertices will be specified by that attribute.
-
-        Besides the usual self-explanatory plotting parameters (C{context},
-        C{bbox}, C{palette}), it accepts the following keyword arguments:
-
-          - C{autocurve}: whether to use curves instead of straight lines for
-            multiple edges on the graph plot. This argument may be C{True}
-            or C{False}; when omitted, C{True} is assumed for graphs with
-            less than 10.000 edges and C{False} otherwise.
-
-          - C{drawer_factory}: a subclass of L{AbstractCairoGraphDrawer}
-            which will be used to draw the graph. You may also provide
-            a function here which takes two arguments: the Cairo context
-            to draw on and a bounding box (an instance of L{BoundingBox}).
-            If this keyword argument is missing, igraph will use the
-            default graph drawer which should be suitable for most purposes.
-            It is safe to omit this keyword argument unless you need to use
-            a specific graph drawer.
-
-          - C{keep_aspect_ratio}: whether to keep the aspect ratio of the layout
-            that igraph calculates to place the nodes. C{True} means that the
-            layout will be scaled proportionally to fit into the bounding box
-            where the graph is to be drawn but the aspect ratio will be kept
-            the same (potentially leaving empty space next to, below or above
-            the graph). C{False} means that the layout will be scaled independently
-            along the X and Y axis in order to fill the entire bounding box.
-            The default is C{False}.
-
-          - C{layout}: the layout to be used. If not an instance of
-            L{Layout}, it will be passed to L{layout} to calculate
-            the layout. Note that if you want a deterministic layout that
-            does not change with every plot, you must either use a
-            deterministic layout function (like L{layout_circle}) or
-            calculate the layout in advance and pass a L{Layout} object here.
-
-          - C{margin}: the top, right, bottom, left margins as a 4-tuple.
-            If it has less than 4 elements or is a single float, the elements
-            will be re-used until the length is at least 4.
-
-          - C{mark_groups}: whether to highlight some of the vertex groups by
-            colored polygons. This argument can be one of the following:
-
-              - C{False}: no groups will be highlighted
-
-              - C{True}: only valid if the object plotted is a
-                L{VertexClustering} or L{VertexCover}. The vertex groups in the
-                clutering or cover will be highlighted such that the i-th
-                group will be colored by the i-th color from the current
-                palette. If used when plotting a graph, it will throw an error.
-
-              - A dict mapping tuples of vertex indices to color names.
-                The given vertex groups will be highlighted by the given
-                colors.
-
-              - A list containing pairs or an iterable yielding pairs, where
-                the first element of each pair is a list of vertex indices and
-                the second element is a color.
-
-              - A L{VertexClustering} or L{VertexCover} instance. The vertex
-                groups in the clustering or cover will be highlighted such that
-                the i-th group will be colored by the i-th color from the
-                current palette.
-
-            In place of lists of vertex indices, you may also use L{VertexSeq}
-            instances.
-
-            In place of color names, you may also use color indices into the
-            current palette. C{None} as a color name will mean that the
-            corresponding group is ignored.
-
-          - C{vertex_size}: size of the vertices. The corresponding vertex
-            attribute is called C{size}. The default is 10. Vertex sizes
-            are measured in the unit of the Cairo context on which igraph
-            is drawing.
-
-          - C{vertex_color}: color of the vertices. The corresponding vertex
-            attribute is C{color}, the default is red.  Colors can be
-            specified either by common X11 color names (see the source
-            code of L{igraph.drawing.colors} for a list of known colors), by
-            3-tuples of floats (ranging between 0 and 255 for the R, G and
-            B components), by CSS-style string specifications (C{#rrggbb})
-            or by integer color indices of the specified palette.
-
-          - C{vertex_frame_color}: color of the frame (i.e. stroke) of the
-            vertices. The corresponding vertex attribute is C{frame_color},
-            the default is black. See C{vertex_color} for the possible ways
-            of specifying a color.
-
-          - C{vertex_frame_width}: the width of the frame (i.e. stroke) of the
-            vertices. The corresponding vertex attribute is C{frame_width}.
-            The default is 1. Vertex frame widths are measured in the unit of the
-            Cairo context on which igraph is drawing.
-
-          - C{vertex_shape}: shape of the vertices. Alternatively it can
-            be specified by the C{shape} vertex attribute. Possibilities
-            are: C{square}, {circle}, {triangle}, {triangle-down} or
-            C{hidden}. See the source code of L{igraph.drawing} for a
-            list of alternative shape names that are also accepted and
-            mapped to these.
-
-          - C{vertex_label}: labels drawn next to the vertices.
-            The corresponding vertex attribute is C{label}.
-
-          - C{vertex_label_dist}: distance of the midpoint of the vertex
-            label from the center of the corresponding vertex.
-            The corresponding vertex attribute is C{label_dist}.
-
-          - C{vertex_label_color}: color of the label. Corresponding
-            vertex attribute: C{label_color}. See C{vertex_color} for
-            color specification syntax.
-
-          - C{vertex_label_size}: font size of the label, specified
-            in the unit of the Cairo context on which we are drawing.
-            Corresponding vertex attribute: C{label_size}.
-
-          - C{vertex_label_angle}: the direction of the line connecting
-            the midpoint of the vertex with the midpoint of the label.
-            This can be used to position the labels relative to the
-            vertices themselves in conjunction with C{vertex_label_dist}.
-            Corresponding vertex attribute: C{label_angle}. The
-            default is C{-math.pi/2}.
-
-          - C{vertex_order}: drawing order of the vertices. This must be
-            a list or tuple containing vertex indices; vertices are then
-            drawn according to this order.
-
-          - C{vertex_order_by}: an alternative way to specify the drawing
-            order of the vertices; this attribute is interpreted as the name
-            of a vertex attribute, and vertices are drawn such that those
-            with a smaller attribute value are drawn first. You may also
-            reverse the order by passing a tuple here; the first element of
-            the tuple should be the name of the attribute, the second element
-            specifies whether the order is reversed (C{True}, C{False},
-            C{"asc"} and C{"desc"} are accepted values).
-
-          - C{edge_color}: color of the edges. The corresponding edge
-            attribute is C{color}, the default is red. See C{vertex_color}
-            for color specification syntax.
-
-          - C{edge_curved}: whether the edges should be curved. Positive
-            numbers correspond to edges curved in a counter-clockwise
-            direction, negative numbers correspond to edges curved in a
-            clockwise direction. Zero represents straight edges. C{True}
-            is interpreted as 0.5, C{False} is interpreted as 0. The
-            default is 0 which makes all the edges straight.
-
-          - C{edge_width}: width of the edges in the default unit of the
-            Cairo context on which we are drawing. The corresponding
-            edge attribute is C{width}, the default is 1.
-
-          - C{edge_arrow_size}: arrow size of the edges. The
-            corresponding edge attribute is C{arrow_size}, the default
-            is 1.
-
-          - C{edge_arrow_width}: width of the arrowhead on the edge. The
-            corresponding edge attribute is C{arrow_width}, the default
-            is 1.
-
-          - C{edge_order}: drawing order of the edges. This must be
-            a list or tuple containing edge indices; edges are then
-            drawn according to this order.
-
-          - C{edge_order_by}: an alternative way to specify the drawing
-            order of the edges; this attribute is interpreted as the name
-            of an edge attribute, and edges are drawn such that those
-            with a smaller attribute value are drawn first. You may also
-            reverse the order by passing a tuple here; the first element of
-            the tuple should be the name of the attribute, the second element
-            specifies whether the order is reversed (C{True}, C{False},
-            C{"asc"} and C{"desc"} are accepted values).
-        """
-        from igraph.drawing import DrawerDirectory
-
-        drawer = kwds.pop(
-            "drawer_factory",
-            DrawerDirectory.resolve(self, backend)(context),
-        )
-        drawer.draw(self, *args, **kwds)
+    __plot__ = _graph_plot
 
     def __str__(self):
         """Returns a string representation of the graph.
@@ -2297,704 +1948,6 @@ class Graph(GraphBase):
 
     # After adjusting something here, don't forget to update the docstring
     # of Graph.layout if necessary!
-
-
-##############################################################
-
-
-class VertexSeq(_VertexSeq):
-    """Class representing a sequence of vertices in the graph.
-
-    This class is most easily accessed by the C{vs} field of the
-    L{Graph} object, which returns an ordered sequence of all vertices in
-    the graph. The vertex sequence can be refined by invoking the
-    L{VertexSeq.select()} method. L{VertexSeq.select()} can also be
-    accessed by simply calling the L{VertexSeq} object.
-
-    An alternative way to create a vertex sequence referring to a given
-    graph is to use the constructor directly:
-
-      >>> g = Graph.Full(3)
-      >>> vs = VertexSeq(g)
-      >>> restricted_vs = VertexSeq(g, [0, 1])
-
-    The individual vertices can be accessed by indexing the vertex sequence
-    object. It can be used as an iterable as well, or even in a list
-    comprehension:
-
-      >>> g=Graph.Full(3)
-      >>> for v in g.vs:
-      ...   v["value"] = v.index ** 2
-      ...
-      >>> [v["value"] ** 0.5 for v in g.vs]
-      [0.0, 1.0, 2.0]
-
-    The vertex set can also be used as a dictionary where the keys are the
-    attribute names. The values corresponding to the keys are the values
-    of the given attribute for every vertex selected by the sequence.
-
-      >>> g=Graph.Full(3)
-      >>> for idx, v in enumerate(g.vs):
-      ...   v["weight"] = idx*(idx+1)
-      ...
-      >>> g.vs["weight"]
-      [0, 2, 6]
-      >>> g.vs.select(1,2)["weight"] = [10, 20]
-      >>> g.vs["weight"]
-      [0, 10, 20]
-
-    If you specify a sequence that is shorter than the number of vertices in
-    the VertexSeq, the sequence is reused:
-
-      >>> g = Graph.Tree(7, 2)
-      >>> g.vs["color"] = ["red", "green"]
-      >>> g.vs["color"]
-      ['red', 'green', 'red', 'green', 'red', 'green', 'red']
-
-    You can even pass a single string or integer, it will be considered as a
-    sequence of length 1:
-
-      >>> g.vs["color"] = "red"
-      >>> g.vs["color"]
-      ['red', 'red', 'red', 'red', 'red', 'red', 'red']
-
-    Some methods of the vertex sequences are simply proxy methods to the
-    corresponding methods in the L{Graph} object. One such example is
-    C{VertexSeq.degree()}:
-
-      >>> g=Graph.Tree(7, 2)
-      >>> g.vs.degree()
-      [2, 3, 3, 1, 1, 1, 1]
-      >>> g.vs.degree() == g.degree()
-      True
-    """
-
-    def attributes(self):
-        """Returns the list of all the vertex attributes in the graph
-        associated to this vertex sequence."""
-        return self.graph.vertex_attributes()
-
-    def find(self, *args, **kwds):
-        """Returns the first vertex of the vertex sequence that matches some
-        criteria.
-
-        The selection criteria are equal to the ones allowed by L{VertexSeq.select}.
-        See L{VertexSeq.select} for more details.
-
-        For instance, to find the first vertex with name C{foo} in graph C{g}:
-
-            >>> g.vs.find(name="foo")            #doctest:+SKIP
-
-        To find an arbitrary isolated vertex:
-
-            >>> g.vs.find(_degree=0)             #doctest:+SKIP
-        """
-        # Shortcut: if "name" is in kwds, there are no positional arguments,
-        # and the specified name is a string, we try that first because that
-        # attribute is indexed. Note that we cannot do this if name is an
-        # integer, because it would then translate to g.vs.select(name), which
-        # searches by _index_ if the argument is an integer
-        if not args:
-            if "name" in kwds:
-                name = kwds.pop("name")
-            elif "name_eq" in kwds:
-                name = kwds.pop("name_eq")
-            else:
-                name = None
-
-            if name is not None:
-                if isinstance(name, str):
-                    args = [name]
-                else:
-                    # put back what we popped
-                    kwds["name"] = name
-
-        if args:
-            # Selecting first based on positional arguments, then checking
-            # the criteria specified by the (remaining) keyword arguments
-            vertex = _VertexSeq.find(self, *args)
-            if not kwds:
-                return vertex
-            vs = self.graph.vs.select(vertex.index)
-        else:
-            vs = self
-
-        # Selecting based on keyword arguments
-        vs = vs.select(**kwds)
-        if vs:
-            return vs[0]
-        raise ValueError("no such vertex")
-
-    def select(self, *args, **kwds):
-        """Selects a subset of the vertex sequence based on some criteria
-
-        The selection criteria can be specified by the positional and the keyword
-        arguments. Positional arguments are always processed before keyword
-        arguments.
-
-          - If the first positional argument is C{None}, an empty sequence is
-            returned.
-
-          - If the first positional argument is a callable object, the object
-            will be called for every vertex in the sequence. If it returns
-            C{True}, the vertex will be included, otherwise it will
-            be excluded.
-
-          - If the first positional argument is an iterable, it must return
-            integers and they will be considered as indices of the current
-            vertex set (NOT the whole vertex set of the graph -- the
-            difference matters when one filters a vertex set that has
-            already been filtered by a previous invocation of
-            L{VertexSeq.select()}. In this case, the indices do not refer
-            directly to the vertices of the graph but to the elements of
-            the filtered vertex sequence.
-
-          - If the first positional argument is an integer, all remaining
-            arguments are expected to be integers. They are considered as
-            indices of the current vertex set again.
-
-        Keyword arguments can be used to filter the vertices based on their
-        attributes. The name of the keyword specifies the name of the attribute
-        and the filtering operator, they should be concatenated by an
-        underscore (C{_}) character. Attribute names can also contain
-        underscores, but operator names don't, so the operator is always the
-        largest trailing substring of the keyword name that does not contain
-        an underscore. Possible operators are:
-
-          - C{eq}: equal to
-
-          - C{ne}: not equal to
-
-          - C{lt}: less than
-
-          - C{gt}: greater than
-
-          - C{le}: less than or equal to
-
-          - C{ge}: greater than or equal to
-
-          - C{in}: checks if the value of an attribute is in a given list
-
-          - C{notin}: checks if the value of an attribute is not in a given
-            list
-
-        For instance, if you want to filter vertices with a numeric C{age}
-        property larger than 200, you have to write:
-
-          >>> g.vs.select(age_gt=200)                   #doctest: +SKIP
-
-        Similarly, to filter vertices whose C{type} is in a list of predefined
-        types:
-
-          >>> list_of_types = ["HR", "Finance", "Management"]
-          >>> g.vs.select(type_in=list_of_types)        #doctest: +SKIP
-
-        If the operator is omitted, it defaults to C{eq}. For instance, the
-        following selector selects vertices whose C{cluster} property equals
-        to 2:
-
-          >>> g.vs.select(cluster=2)                    #doctest: +SKIP
-
-        In the case of an unknown operator, it is assumed that the
-        recognized operator is part of the attribute name and the actual
-        operator is C{eq}.
-
-        Attribute names inferred from keyword arguments are treated specially
-        if they start with an underscore (C{_}). These are not real attributes
-        but refer to specific properties of the vertices, e.g., its degree.
-        The rule is as follows: if an attribute name starts with an underscore,
-        the rest of the name is interpreted as a method of the L{Graph} object.
-        This method is called with the vertex sequence as its first argument
-        (all others left at default values) and vertices are filtered
-        according to the value returned by the method. For instance, if you
-        want to exclude isolated vertices:
-
-          >>> g = Graph.Famous("zachary")
-          >>> non_isolated = g.vs.select(_degree_gt=0)
-
-        For properties that take a long time to be computed (e.g., betweenness
-        centrality for large graphs), it is advised to calculate the values
-        in advance and store it in a graph attribute. The same applies when
-        you are selecting based on the same property more than once in the
-        same C{select()} call to avoid calculating it twice unnecessarily.
-        For instance, the following would calculate betweenness centralities
-        twice:
-
-          >>> edges = g.vs.select(_betweenness_gt=10, _betweenness_lt=30)
-
-        It is advised to use this instead:
-
-          >>> g.vs["bs"] = g.betweenness()
-          >>> edges = g.vs.select(bs_gt=10, bs_lt=30)
-
-        @return: the new, filtered vertex sequence"""
-        vs = _VertexSeq.select(self, *args)
-
-        operators = {
-            "lt": operator.lt,
-            "gt": operator.gt,
-            "le": operator.le,
-            "ge": operator.ge,
-            "eq": operator.eq,
-            "ne": operator.ne,
-            "in": lambda a, b: a in b,
-            "notin": lambda a, b: a not in b,
-        }
-        for keyword, value in kwds.items():
-            if "_" not in keyword or keyword.rindex("_") == 0:
-                keyword = keyword + "_eq"
-            attr, _, op = keyword.rpartition("_")
-            try:
-                func = operators[op]
-            except KeyError:
-                # No such operator, assume that it's part of the attribute name
-                attr, op, func = keyword, "eq", operators["eq"]
-
-            if attr[0] == "_":
-                # Method call, not an attribute
-                values = getattr(vs.graph, attr[1:])(vs)
-            else:
-                values = vs[attr]
-            filtered_idxs = [i for i, v in enumerate(values) if func(v, value)]
-            vs = vs.select(filtered_idxs)
-
-        return vs
-
-    def __call__(self, *args, **kwds):
-        """Shorthand notation to select()
-
-        This method simply passes all its arguments to L{VertexSeq.select()}.
-        """
-        return self.select(*args, **kwds)
-
-
-##############################################################
-
-
-class EdgeSeq(_EdgeSeq):
-    """Class representing a sequence of edges in the graph.
-
-    This class is most easily accessed by the C{es} field of the
-    L{Graph} object, which returns an ordered sequence of all edges in
-    the graph. The edge sequence can be refined by invoking the
-    L{EdgeSeq.select()} method. L{EdgeSeq.select()} can also be
-    accessed by simply calling the L{EdgeSeq} object.
-
-    An alternative way to create an edge sequence referring to a given
-    graph is to use the constructor directly:
-
-      >>> g = Graph.Full(3)
-      >>> es = EdgeSeq(g)
-      >>> restricted_es = EdgeSeq(g, [0, 1])
-
-    The individual edges can be accessed by indexing the edge sequence
-    object. It can be used as an iterable as well, or even in a list
-    comprehension:
-
-      >>> g=Graph.Full(3)
-      >>> for e in g.es:
-      ...   print(e.tuple)
-      ...
-      (0, 1)
-      (0, 2)
-      (1, 2)
-      >>> [max(e.tuple) for e in g.es]
-      [1, 2, 2]
-
-    The edge sequence can also be used as a dictionary where the keys are the
-    attribute names. The values corresponding to the keys are the values
-    of the given attribute of every edge in the graph:
-
-      >>> g=Graph.Full(3)
-      >>> for idx, e in enumerate(g.es):
-      ...   e["weight"] = idx*(idx+1)
-      ...
-      >>> g.es["weight"]
-      [0, 2, 6]
-      >>> g.es["weight"] = range(3)
-      >>> g.es["weight"]
-      [0, 1, 2]
-
-    If you specify a sequence that is shorter than the number of edges in
-    the EdgeSeq, the sequence is reused:
-
-      >>> g = Graph.Tree(7, 2)
-      >>> g.es["color"] = ["red", "green"]
-      >>> g.es["color"]
-      ['red', 'green', 'red', 'green', 'red', 'green']
-
-    You can even pass a single string or integer, it will be considered as a
-    sequence of length 1:
-
-      >>> g.es["color"] = "red"
-      >>> g.es["color"]
-      ['red', 'red', 'red', 'red', 'red', 'red']
-
-    Some methods of the edge sequences are simply proxy methods to the
-    corresponding methods in the L{Graph} object. One such example is
-    C{EdgeSeq.is_multiple()}:
-
-      >>> g=Graph(3, [(0,1), (1,0), (1,2)])
-      >>> g.es.is_multiple()
-      [False, True, False]
-      >>> g.es.is_multiple() == g.is_multiple()
-      True
-    """
-
-    def attributes(self):
-        """Returns the list of all the edge attributes in the graph
-        associated to this edge sequence."""
-        return self.graph.edge_attributes()
-
-    def find(self, *args, **kwds):
-        """Returns the first edge of the edge sequence that matches some
-        criteria.
-
-        The selection criteria are equal to the ones allowed by L{VertexSeq.select}.
-        See L{VertexSeq.select} for more details.
-
-        For instance, to find the first edge with weight larger than 5 in graph C{g}:
-
-            >>> g.es.find(weight_gt=5)           #doctest:+SKIP
-        """
-        if args:
-            # Selecting first based on positional arguments, then checking
-            # the criteria specified by the keyword arguments
-            edge = _EdgeSeq.find(self, *args)
-            if not kwds:
-                return edge
-            es = self.graph.es.select(edge.index)
-        else:
-            es = self
-
-        # Selecting based on positional arguments
-        es = es.select(**kwds)
-        if es:
-            return es[0]
-        raise ValueError("no such edge")
-
-    def select(self, *args, **kwds):
-        """Selects a subset of the edge sequence based on some criteria
-
-        The selection criteria can be specified by the positional and the
-        keyword arguments. Positional arguments are always processed before
-        keyword arguments.
-
-          - If the first positional argument is C{None}, an empty sequence is
-            returned.
-
-          - If the first positional argument is a callable object, the object
-            will be called for every edge in the sequence. If it returns
-            C{True}, the edge will be included, otherwise it will
-            be excluded.
-
-          - If the first positional argument is an iterable, it must return
-            integers and they will be considered as indices of the current
-            edge set (NOT the whole edge set of the graph -- the
-            difference matters when one filters an edge set that has
-            already been filtered by a previous invocation of
-            L{EdgeSeq.select()}. In this case, the indices do not refer
-            directly to the edges of the graph but to the elements of
-            the filtered edge sequence.
-
-          - If the first positional argument is an integer, all remaining
-            arguments are expected to be integers. They are considered as
-            indices of the current edge set again.
-
-        Keyword arguments can be used to filter the edges based on their
-        attributes and properties. The name of the keyword specifies the name
-        of the attribute and the filtering operator, they should be
-        concatenated by an underscore (C{_}) character. Attribute names can
-        also contain underscores, but operator names don't, so the operator is
-        always the largest trailing substring of the keyword name that does not
-        contain an underscore. Possible operators are:
-
-          - C{eq}: equal to
-
-          - C{ne}: not equal to
-
-          - C{lt}: less than
-
-          - C{gt}: greater than
-
-          - C{le}: less than or equal to
-
-          - C{ge}: greater than or equal to
-
-          - C{in}: checks if the value of an attribute is in a given list
-
-          - C{notin}: checks if the value of an attribute is not in a given
-            list
-
-        For instance, if you want to filter edges with a numeric C{weight}
-        property larger than 50, you have to write:
-
-          >>> g.es.select(weight_gt=50)            #doctest: +SKIP
-
-        Similarly, to filter edges whose C{type} is in a list of predefined
-        types:
-
-          >>> list_of_types = ["inhibitory", "excitatory"]
-          >>> g.es.select(type_in=list_of_types)   #doctest: +SKIP
-
-        If the operator is omitted, it defaults to C{eq}. For instance, the
-        following selector selects edges whose C{type} property is
-        C{intracluster}:
-
-          >>> g.es.select(type="intracluster")     #doctest: +SKIP
-
-        In the case of an unknown operator, it is assumed that the
-        recognized operator is part of the attribute name and the actual
-        operator is C{eq}.
-
-        Keyword arguments are treated specially if they start with an
-        underscore (C{_}). These are not real attributes but refer to specific
-        properties of the edges, e.g., their centrality.  The rules are as
-        follows:
-
-          1. C{_source} or {_from} means the source vertex of an edge. For
-             undirected graphs, only the C{eq} operator is supported and it
-             is treated as {_incident} (since undirected graphs have no notion
-             of edge directionality).
-
-          2. C{_target} or {_to} means the target vertex of an edge. For
-             undirected graphs, only the C{eq} operator is supported and it
-             is treated as {_incident} (since undirected graphs have no notion
-             of edge directionality).
-
-          3. C{_within} ignores the operator and checks whether both endpoints
-             of the edge lie within a specified set.
-
-          4. C{_between} ignores the operator and checks whether I{one}
-             endpoint of the edge lies within a specified set and the I{other}
-             endpoint lies within another specified set. The two sets must be
-             given as a tuple.
-
-          5. C{_incident} ignores the operator and checks whether the edge is
-             incident on a specific vertex or a set of vertices.
-
-          6. Otherwise, the rest of the name is interpreted as a method of the
-             L{Graph} object. This method is called with the edge sequence as
-             its first argument (all others left at default values) and edges
-             are filtered according to the value returned by the method.
-
-        For instance, if you want to exclude edges with a betweenness
-        centrality less than 2:
-
-          >>> g = Graph.Famous("zachary")
-          >>> excl = g.es.select(_edge_betweenness_ge = 2)
-
-        To select edges originating from vertices 2 and 4:
-
-          >>> edges = g.es.select(_source_in = [2, 4])
-
-        To select edges lying entirely within the subgraph spanned by vertices
-        2, 3, 4 and 7:
-
-          >>> edges = g.es.select(_within = [2, 3, 4, 7])
-
-        To select edges with one endpoint in the vertex set containing vertices
-        2, 3, 4 and 7 and the other endpoint in the vertex set containing
-        vertices 8 and 9:
-
-          >>> edges = g.es.select(_between = ([2, 3, 4, 7], [8, 9]))
-
-        For properties that take a long time to be computed (e.g., betweenness
-        centrality for large graphs), it is advised to calculate the values
-        in advance and store it in a graph attribute. The same applies when
-        you are selecting based on the same property more than once in the
-        same C{select()} call to avoid calculating it twice unnecessarily.
-        For instance, the following would calculate betweenness centralities
-        twice:
-
-          >>> edges = g.es.select(_edge_betweenness_gt=10,       # doctest:+SKIP
-          ...                     _edge_betweenness_lt=30)
-
-        It is advised to use this instead:
-
-          >>> g.es["bs"] = g.edge_betweenness()
-          >>> edges = g.es.select(bs_gt=10, bs_lt=30)
-
-        @return: the new, filtered edge sequence
-        """
-        es = _EdgeSeq.select(self, *args)
-        is_directed = self.graph.is_directed()
-
-        def _ensure_set(value):
-            if isinstance(value, VertexSeq):
-                value = set(v.index for v in value)
-            elif not isinstance(value, (set, frozenset)):
-                value = set(value)
-            return value
-
-        operators = {
-            "lt": operator.lt,
-            "gt": operator.gt,
-            "le": operator.le,
-            "ge": operator.ge,
-            "eq": operator.eq,
-            "ne": operator.ne,
-            "in": lambda a, b: a in b,
-            "notin": lambda a, b: a not in b,
-        }
-
-        # TODO(ntamas): some keyword arguments should be prioritized over
-        # others; for instance, we have optimized code paths for _source and
-        # _target in directed and undirected graphs if es.is_all() is True;
-        # these should be executed first. This matters only if there are
-        # multiple keyword arguments and es.is_all() is True.
-
-        for keyword, value in kwds.items():
-            if "_" not in keyword or keyword.rindex("_") == 0:
-                keyword = keyword + "_eq"
-            pos = keyword.rindex("_")
-            attr, op = keyword[0:pos], keyword[pos + 1 :]
-            try:
-                func = operators[op]
-            except KeyError:
-                # No such operator, assume that it's part of the attribute name
-                attr, op, func = keyword, "eq", operators["eq"]
-
-            if attr[0] == "_":
-                if attr in ("_source", "_from", "_target", "_to") and not is_directed:
-                    if op not in ("eq", "in"):
-                        raise RuntimeError("unsupported for undirected graphs")
-
-                    # translate to _incident to avoid confusion
-                    attr = "_incident"
-                    if func == operators["eq"]:
-                        if hasattr(value, "__iter__") and not isinstance(value, str):
-                            value = set(value)
-                        else:
-                            value = set([value])
-
-                if attr in ("_source", "_from"):
-                    if es.is_all() and op == "eq":
-                        # shortcut here: use .incident() as it is much faster
-                        filtered_idxs = sorted(es.graph.incident(value, mode="out"))
-                        func = None
-                        # TODO(ntamas): there are more possibilities; we could
-                        # optimize "ne", "in" and "notin" in similar ways
-                    else:
-                        values = [e.source for e in es]
-                        if op == "in" or op == "notin":
-                            value = _ensure_set(value)
-
-                elif attr in ("_target", "_to"):
-                    if es.is_all() and op == "eq":
-                        # shortcut here: use .incident() as it is much faster
-                        filtered_idxs = sorted(es.graph.incident(value, mode="in"))
-                        func = None
-                        # TODO(ntamas): there are more possibilities; we could
-                        # optimize "ne", "in" and "notin" in similar ways
-                    else:
-                        values = [e.target for e in es]
-                        if op == "in" or op == "notin":
-                            value = _ensure_set(value)
-
-                elif attr == "_incident":
-                    func = None  # ignoring function, filtering here
-                    value = _ensure_set(value)
-
-                    # Fetch all the edges that are incident on at least one of
-                    # the vertices specified
-                    candidates = set()
-                    for v in value:
-                        candidates.update(es.graph.incident(v))
-
-                    if not es.is_all():
-                        # Find those that are in the current edge sequence
-                        filtered_idxs = [
-                            i for i, e in enumerate(es) if e.index in candidates
-                        ]
-                    else:
-                        # We are done, the filtered indexes are in the candidates set
-                        filtered_idxs = sorted(candidates)
-
-                elif attr == "_within":
-                    func = None  # ignoring function, filtering here
-                    value = _ensure_set(value)
-
-                    # Fetch all the edges that are incident on at least one of
-                    # the vertices specified
-                    candidates = set()
-                    for v in value:
-                        candidates.update(es.graph.incident(v))
-
-                    if not es.is_all():
-                        # Find those where both endpoints are OK
-                        filtered_idxs = [
-                            i
-                            for i, e in enumerate(es)
-                            if e.index in candidates
-                            and e.source in value
-                            and e.target in value
-                        ]
-                    else:
-                        # Optimized version when the edge sequence contains all
-                        # the edges exactly once in increasing order of edge IDs
-                        filtered_idxs = [
-                            i
-                            for i in candidates
-                            if es[i].source in value and es[i].target in value
-                        ]
-
-                elif attr == "_between":
-                    if len(value) != 2:
-                        raise ValueError(
-                            "_between selector requires two vertex ID lists"
-                        )
-                    func = None  # ignoring function, filtering here
-                    set1 = _ensure_set(value[0])
-                    set2 = _ensure_set(value[1])
-
-                    # Fetch all the edges that are incident on at least one of
-                    # the vertices specified
-                    candidates = set()
-                    for v in set1:
-                        candidates.update(es.graph.incident(v))
-                    for v in set2:
-                        candidates.update(es.graph.incident(v))
-
-                    if not es.is_all():
-                        # Find those where both endpoints are OK
-                        filtered_idxs = [
-                            i
-                            for i, e in enumerate(es)
-                            if (e.source in set1 and e.target in set2)
-                            or (e.target in set1 and e.source in set2)
-                        ]
-                    else:
-                        # Optimized version when the edge sequence contains all
-                        # the edges exactly once in increasing order of edge IDs
-                        filtered_idxs = [
-                            i
-                            for i in candidates
-                            if (es[i].source in set1 and es[i].target in set2)
-                            or (es[i].target in set1 and es[i].source in set2)
-                        ]
-
-                else:
-                    # Method call, not an attribute
-                    values = getattr(es.graph, attr[1:])(es)
-            else:
-                values = es[attr]
-
-            # If we have a function to apply on the values, do that; otherwise
-            # we assume that filtered_idxs has already been calculated.
-            if func is not None:
-                filtered_idxs = [i for i, v in enumerate(values) if func(v, value)]
-
-            es = es.select(filtered_idxs)
-
-        return es
-
-    def __call__(self, *args, **kwds):
-        """Shorthand notation to select()
-
-        This method simply passes all its arguments to L{EdgeSeq.select()}.
-        """
-        return self.select(*args, **kwds)
 
 
 ##############################################################
@@ -3172,69 +2125,6 @@ Graph.layout_sphere = _3d_version_for(Graph.layout_circle)
 ##############################################################
 
 
-def autocurve(graph, attribute="curved", default=0):
-    """Calculates curvature values for each of the edges in the graph to make
-    sure that multiple edges are shown properly on a graph plot.
-
-    This function checks the multiplicity of each edge in the graph and
-    assigns curvature values (numbers between -1 and 1, corresponding to
-    CCW (-1), straight (0) and CW (1) curved edges) to them. The assigned
-    values are either stored in an edge attribute or returned as a list,
-    depending on the value of the I{attribute} argument.
-
-    @param graph: the graph on which the calculation will be run
-    @param attribute: the name of the edge attribute to save the curvature
-      values to. The default value is C{curved}, which is the name of the
-      edge attribute the default graph plotter checks to decide whether an
-      edge should be curved on the plot or not. If I{attribute} is C{None},
-      the result will not be stored.
-    @param default: the default curvature for single edges. Zero means that
-      single edges will be straight. If you want single edges to be curved
-      as well, try passing 0.5 or -0.5 here.
-    @return: the list of curvature values if I{attribute} is C{None},
-      otherwise C{None}.
-    """
-
-    # The following loop could be re-written in C if it turns out to be a
-    # bottleneck. Unfortunately we cannot use Graph.count_multiple() here
-    # because we have to ignore edge directions.
-    multiplicities = defaultdict(list)
-    for edge in graph.es:
-        u, v = edge.tuple
-        if u > v:
-            multiplicities[v, u].append(edge.index)
-        else:
-            multiplicities[u, v].append(edge.index)
-
-    result = [default] * graph.ecount()
-    for eids in multiplicities.values():
-        # Is it a single edge?
-        if len(eids) < 2:
-            continue
-
-        if len(eids) % 2 == 1:
-            # Odd number of edges; the last will be straight
-            result[eids.pop()] = 0
-
-        # Arrange the remaining edges
-        curve = 2.0 / (len(eids) + 2)
-        dcurve, sign = curve, 1
-        for idx, eid in enumerate(eids):
-            edge = graph.es[eid]
-            if edge.source > edge.target:
-                result[eid] = -sign * curve
-            else:
-                result[eid] = sign * curve
-            if idx % 2 == 1:
-                curve += dcurve
-            sign *= -1
-
-    if attribute is None:
-        return result
-
-    graph.es[attribute] = result
-
-
 def get_include():
     """Returns the folder that contains the C API headers of the Python
     interface of igraph."""
@@ -3330,4 +2220,6 @@ del (
     _community_walktrap,
     _k_core,
     _community_leiden,
+    _graph_plot,
+    _operator_method_registry,
 )
