+++ src/igraph/operators/functions.py
